% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treas.R
\name{treas_auctioned}
\alias{treas_auctioned}
\title{List recently auctioned Treasury securities}
\usage{
treas_auctioned(days = 7, limit = 50)
}
\arguments{
\item{days}{Number of days to look back (default 0 = today, 7 = last week)}
\item{limit}{Maximum number of results (default 50, max 250)}
}
\value{
tibble: cusip, security_type, security_term, issue_date,
  maturity_date, interest_rate, auction_date, offering_amount,
  high_yield, bid_to_cover_ratio, reopening
}
\description{
List recently auctioned Treasury securities
}
