% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treas.R
\name{treas_debt_current}
\alias{treas_debt_current}
\title{Get current U.S. national debt totals}
\usage{
treas_debt_current()
}
\value{
tibble: effective_date, public_debt, government_holdings, total_debt
}
\description{
Get current U.S. national debt totals
}
