% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treas.R
\name{treas_security}
\alias{treas_security}
\title{Get details for a specific Treasury security by CUSIP}
\usage{
treas_security(cusip)
}
\arguments{
\item{cusip}{CUSIP identifier (e.g., "912810UR7")}
}
\value{
tibble: single row with security details
}
\description{
Get details for a specific Treasury security by CUSIP
}
