% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treas.R
\name{treas_upcoming}
\alias{treas_upcoming}
\title{List upcoming Treasury security auctions}
\usage{
treas_upcoming(limit = 50)
}
\arguments{
\item{limit}{Maximum number of results (default 50, max 250)}
}
\value{
tibble: cusip, security_type, security_term, issue_date,
  maturity_date, interest_rate, auction_date, offering_amount,
  high_yield, bid_to_cover_ratio, reopening
}
\description{
List upcoming Treasury security auctions
}
